# src/models/req/ReqEmployee.ts
export interface ReqEmployee {
    firstName?: string,
    lastName?: string,
    status?: boolean,
    salary?: number,
    date_of_joining: Date,
    username?: string,
    password?: string,
    // department?: number,
    // branch?: number,
    // company?: number,
    role?: number,
    skills?: number[]
}