# src/tsoa-auto/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ResCompany": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"logo_url": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResBranch": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"company": {
						"$ref": "#/components/schemas/ResCompany"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResDepartment": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"branch": {
						"$ref": "#/components/schemas/ResBranch"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResRole": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"roleName": {
						"type": "string"
					},
					"roleDescription": {
						"type": "string"
					},
					"permission": {
						"items": {
							"$ref": "#/components/schemas/ResPermission"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResPermission": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"permission_name": {
						"type": "string"
					},
					"permission_description": {
						"type": "string"
					},
					"role": {
						"items": {
							"$ref": "#/components/schemas/ResRole"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResSkill": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResEmployee": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"username": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					},
					"department": {
						"$ref": "#/components/schemas/ResDepartment"
					},
					"branch": {
						"$ref": "#/components/schemas/ResBranch"
					},
					"company": {
						"$ref": "#/components/schemas/ResCompany"
					},
					"role": {
						"$ref": "#/components/schemas/ResRole"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/ResSkill"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ResError": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ReqEmployee": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"date_of_joining": {
						"type": "string",
						"format": "date-time"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"type": "number",
						"format": "double"
					},
					"skills": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"date_of_joining"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserLoginCompany": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"logo_url": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"logo_url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserLoginRole": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"role_name": {
						"type": "string"
					},
					"role_description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"role_name",
					"role_description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResUserLogin": {
				"properties": {
					"name": {
						"type": "string"
					},
					"company": {
						"$ref": "#/components/schemas/UserLoginCompany"
					},
					"branch": {
						"$ref": "#/components/schemas/ResBranch"
					},
					"department": {
						"$ref": "#/components/schemas/ResDepartment"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/ResPermission"
						},
						"type": "array"
					},
					"role": {
						"$ref": "#/components/schemas/UserLoginRole"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"company",
					"branch",
					"department",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmpLogin": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResSuccess": {
				"properties": {
					"result": {
						"type": "string"
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Api-Token": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "WorkHive APIs",
		"version": "0.0.1",
		"description": "WorkhiveServer",
		"contact": {}
	},
	"paths": {
...
		"/employee/{employeeId}": {
			"get": {
				"operationId": "GetOneEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ResEmployee"
										},
										{
											"$ref": "#/components/schemas/ResError"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResSuccess"
								}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/employee": {
			"get": {
				"operationId": "GetAllEmployeeBranch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ResEmployee"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "gets all employee in branch",
				"tags": [
					"Employee"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "SaveEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResEmployee"
								}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqEmployee"
							}
						}
					}
				}
			}
		},
		"/employee/login": {
			"post": {
				"operationId": "UserLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResUserLogin"
								}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmpLogin"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}